[ feat ] 확장자 전부 변경, ShuffledCard 변경
