+++ TS-cardGame/ts-card/src/components/Modal.tsx
