+++ TS-cardGame/ts-card/src/components/ResetBtn.tsx
