+++ TS-cardGame/ts-card/src/components/SingleCard.tsx
