+++ TS-cardGame/ts-card/tsconfig.json
@@ -5,7 +5,7 @@
     "lib": ["ES2020", "DOM", "DOM.Iterable"],
     "module": "ESNext",
     "skipLibCheck": true,
-    "allowJs": true,
+    // "allowJs": true,
 
     /* Bundler mode */
     "moduleResolution": "bundler",
